+++ src/views/pages/PersistentVolumeManagement.vue
@@ -16,58 +16,14 @@ import { getHttpBaseUrl } from '@/vendor/utils.js'
 import axios from 'axios'
 import { useAuthStore } from '@/store/auth.js'
 import Badge from '@/views/components/Badge.vue'
+import CreatePersistentVolumeModal from '@/views/partials/CreatePersistentVolumeModal.vue'
 
 const toast = useToast()
 const authStore = useAuthStore()
-const isModalOpen = ref(false)
-const openModal = () => {
-  isModalOpen.value = true
-}
-const closeModal = () => {
-  isModalOpen.value = false
-}
 
 // Create persistent volume
-const newPersistentVolumeDetails = reactive({
-  name: '',
-  type: 'local',
-  nfsConfig: {
-    host: '',
-    path: '',
-    version: 4
-  }
-})
-
-const {
-  mutate: registerPersistentVolume,
-  loading: isDomainRegistering,
-  onDone: onDomainRegisterSuccess,
-  onError: onDomainRegisterFail
-} = useMutation(
-  gql`
-    mutation ($input: PersistentVolumeInput!) {
-      createPersistentVolume(input: $input) {
-        id
-        name
-      }
-    }
-  `,
-  {
-    variables: {
-      input: newPersistentVolumeDetails
-    }
-  }
-)
-
-onDomainRegisterSuccess(() => {
-  closeModal()
-  newPersistentVolumeDetails.name = ''
-  refetchPersistentVolumes()
-})
-
-onDomainRegisterFail((err) => {
-  toast.error(err.message)
-})
+const createPersistentVolumeModalRef = ref(null)
+const openCreatePersistentVolumeModal = computed(() => createPersistentVolumeModalRef.value?.openModal ?? (() => {}))
 
 // Delete persistent volume
 const {
@@ -227,93 +183,16 @@ const showDetails = (volume) => {
       :close-drawer="closeBackupDrawer"
       :persistent-volume-id="selectedPersistentVolumeId"
       :persistent-volume-name="selectedPersistentVolumeName" />
+
     <!-- Drawer for persistent volume restores -->
     <PersistentVolumeRestores
       :is-drawer-open="isRestoreDrawerOpen"
       :close-drawer="closeRestoreDrawer"
       :persistent-volume-id="selectedPersistentVolumeId"
       :persistent-volume-name="selectedPersistentVolumeName" />
+
     <!-- Modal for create persistent volumes -->
-    <ModalDialog :close-modal="closeModal" :is-open="isModalOpen">
-      <template v-slot:header>Add New Persistent Volume</template>
-      <template v-slot:body>
-        Enter a unique name for the persistent volume.
-        <form @submit.prevent="">
-          <!--  Name Field   -->
-          <div class="mt-4">
-            <label class="block text-sm font-medium text-gray-700" for="name"> Persistent Volume </label>
-            <div class="mt-1">
-              <input
-                id="name"
-                v-model="newPersistentVolumeDetails.name"
-                autocomplete="off"
-                class="block w-full rounded-md border-gray-300 shadow-sm focus:border-primary-500 focus:ring-primary-500 sm:text-sm"
-                name="name"
-                placeholder="Name of persistent volume"
-                type="text" />
-            </div>
-          </div>
-          <!--    Type Field      -->
-          <div class="mt-2">
-            <label class="block text-sm font-medium text-gray-700">Type</label>
-            <select
-              v-model="newPersistentVolumeDetails.type"
-              class="mt-2 block w-full rounded-md border-gray-300 shadow-sm focus:border-primary-500 focus:ring-primary-500 sm:text-sm">
-              <option value="local">Local</option>
-              <option value="nfs">NFS</option>
-            </select>
-          </div>
-          <!--   NFS Server Host    -->
-          <div class="mt-2" v-if="newPersistentVolumeDetails.type === 'nfs'">
-            <label class="block text-sm font-medium text-gray-700">NFS Server Host</label>
-            <div class="mt-1">
-              <input
-                v-model="newPersistentVolumeDetails.nfsConfig.host"
-                autocomplete="off"
-                class="block w-full rounded-md border-gray-300 shadow-sm focus:border-primary-500 focus:ring-primary-500 sm:text-sm"
-                placeholder="NFS Server Host"
-                type="text" />
-            </div>
-            <p class="mt-2 text-sm text-gray-500">
-              Example:
-              <span class="text-gray-700"> nfs-server.example.com </span>
-            </p>
-          </div>
-          <!--    NFS Share Path      -->
-          <div class="mt-2" v-if="newPersistentVolumeDetails.type === 'nfs'">
-            <label class="block text-sm font-medium text-gray-700">NFS Share Path</label>
-            <div class="mt-1">
-              <input
-                v-model="newPersistentVolumeDetails.nfsConfig.path"
-                autocomplete="off"
-                class="block w-full rounded-md border-gray-300 shadow-sm focus:border-primary-500 focus:ring-primary-500 sm:text-sm"
-                placeholder="NFS Share Path"
-                type="text" />
-            </div>
-            <p class="mt-2 text-sm text-gray-500">
-              Example:
-              <span class="text-gray-700"> /mnt/nfs_share </span>
-            </p>
-          </div>
-          <!--  Version -->
-          <div class="mt-2" v-if="newPersistentVolumeDetails.type === 'nfs'">
-            <label class="block text-sm font-medium text-gray-700">NFS Version</label>
-            <select
-              v-model="newPersistentVolumeDetails.nfsConfig.version"
-              class="mt-2 block w-full rounded-md border-gray-300 shadow-sm focus:border-primary-500 focus:ring-primary-500 sm:text-sm">
-              <option value="2">NFS v3</option>
-              <option value="3">NFS v3</option>
-              <option value="4">NFS v4</option>
-            </select>
-          </div>
-        </form>
-      </template>
-      <template v-slot:footer>
-        <FilledButton :click="registerPersistentVolume" :loading="isDomainRegistering" type="primary"
-          >Register
-        </FilledButton>
-      </template>
-    </ModalDialog>
+    <CreatePersistentVolumeModal ref="createPersistentVolumeModalRef" :callback-on-create="refetchPersistentVolumes" />
 
     <!--    Modal for create restore -->
     <ModalDialog :close-modal="closeRestoreNowModal" :is-open="isRestoreNowModalOpen" non-cancelable>
@@ -401,7 +280,7 @@ const showDetails = (volume) => {
       <template v-slot:title>Persistent Volume</template>
       <template v-slot:subtitle>Manage Persistent Volume</template>
       <template v-slot:buttons>
-        <FilledButton :click="openModal" type="primary">Add New</FilledButton>
+        <FilledButton :click="openCreatePersistentVolumeModal" type="primary">Add New</FilledButton>
       </template>
     </PageBar>
